feat: created the form input todo and badge
Created form input todo element which is wrapped in the FormInputTodo component as a children's.

And also created the `categories` array (badge children), which contains all the types category in the app and returns the Badge component.

Then pass the isActive property to the Badge component, The value of isActive is based on the selectedCategory state. If the current selectedCategory state matches the children's badge then the value is true.