feat: allow user to add a new todo
Import useDispatch hook and addTodo action.

Create handleAddNewTodo to handle add todo item to the state using addTodo action if the value of the newTodoValue state is > 2. If the value of newTodoValue state is < 2 then add the 'animate-shaking' classname to the input new todo field.

Refactor the section.motion the the section